FIx the blocking read bug
After disassembling Tetris and dowloanding ghidra
Found out that the ioport removed the 2 bits from the stat register.
Removed the line in read.
Also moved the ppu to work before the interrupt handler,
cause the cpu could request an interrupt and since one of the interrupts
depends on the ppu the ppu needs to work before the interrupt handler.