Fix the wave channel
The problem was that the channel did not had a volume.
So any use of the current_volume field was inaccurate,
since the sample producer hadnles the volume the channel iteself
is setting the current_volume to 1.
this way if the volume is 0, the sample producer will return 0.

I think this didnt work before this commit cause the current_volume only
loads on trigger, so every change after will not take affect in a
channel with no volume envelope