Fix sprite fetcher
The fifo is 8 pxiels length.
If it contains pixels from previous sprite the new sprite first
number of pixels in the fifo are discarded and not pushed to the fifo.
This way sprite whos ovelapping wont override each other
and the sprite with the lowest x value (higher priority) will be drawn.