Fix some bugs causing the program to panic
mbc1 - casting to u16 when the program length
in mbc1 carts can be greater than u16.
read or write to the ram
when there is not ram read or writes are ignored
Add support for ignoring and not crashing

gb_ppu - when calculating the next ppu state
change did not count the vblank state when the
counter get greater than 114 m_cycles.
Now Im using modolu of 114 on the cycles to handle this

sprite_fetcher - a bug where the flipx calculation
was off by the amount of pixels needs to skip
causing panic.
It seems to be unecessary so I reomved it.